requirements.txt
streamlit
transformers>=4.22.1
torch
Pillow
requests
diffusers
accelerate